# base/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Học Tiếng Anh",
        "description": "Tài liệu API cho ứng dụng học tiếng Anh.",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập người dùng",
                "operationId": "fb66b153966d9fac2e999136ef0ade67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công"
                    },
                    "401": {
                        "description": "Thông tin đăng nhập không hợp lệ"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký người dùng mới",
                "operationId": "fb835c97347e657e2a627790c4c5817a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Người dùng đã được tạo thành công"
                    },
                    "422": {
                        "description": "Lỗi xác thực"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng xuất người dùng",
                "operationId": "307ed0512855922b2280792376bb2095",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Thay đổi mật khẩu người dùng",
                "operationId": "bd4dd230facd62e300c1bed2cc7f512f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thay đổi mật khẩu thành công"
                    },
                    "400": {
                        "description": "Lỗi xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}